refactor: remove ActionButton component and update imports in BookOwnerCard
